mysite/requirements.txt
Django==2.2.6
django-ckeditor==5.8.0
django-js-asset==1.2.2
pytz==2019.2
sqlparse==0.3.0
virtualenv==16.7.5
virtualenvwrapper-win==1.2.5
